+++ src/SimulateTurn/Actions.hpp
@@ -4,6 +4,8 @@
 
 namespace pokesim {
 struct SideDecision;
+struct ActionQueue;
 
 void resolveDecision(types::handle sideHandle, const SideDecision& sideDecision);
+void speedSort(types::handle handle, ActionQueue& actionQueue);
 }  // namespace pokesim
