--- src/SimulateTurn/Actions/SpeedSort.hpp
@@ -1,8 +1,0 @@
-#pragma once
-
-#include <Types/Entity.hpp>
-
-namespace pokesim {
-struct ActionQueue;
-void speedSort(types::handle handle, ActionQueue& actionQueue);
-}  // namespace pokesim
