--- src/SimulateTurn/Actions/headers.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "Decisions.hpp"
-#include "ResolveDecision.hpp"
-#include "SpeedSort.hpp"
